fix: move sigining test
